requirements.txt
os
PyPDF2
docx
re
langchain-ollam
json
reportlab
fastapi